+++ src/maths/Vector3D.hpp
@@ -50,7 +50,7 @@ namespace Math {
         // Methods
         [[nodiscard]] double dot(const Vector3D &vector) const;
         [[nodiscard]] Vector3D cross(const Vector3D &other) const;
-        [[nodiscard]] float length() const;
+        [[nodiscard]] double length() const;
 
         // Variables
         double _x;
